Fix some minor fuzz-testing findings

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix addressing multiple issues revealed by fuzz testing.

#### PR Summary
This pull request improves input validation, pointer arithmetic, and memory cleanup in SPDM response processing and error logging.
- `core/spdm/spdm_commands.c`: Refined length checks, pointer calculations, and algorithm structure table iteration for negotiate algorithms and measurements responses.
- `core/spdm/cmd_interface_spdm.c`: Adjusted error logging by enforcing proper type casting in bit-shift operations.
- `core/spdm/spdm_commands.h`: Corrected pointer arithmetic in the algorithm table entry macro using the appropriate type cast.
- `core/manifest/cfm/cfm_flash.c`: Added a reset for the manifest's check count during cleanup.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

----
#### AI description  (iteration 3)
#### PR Classification
Bug fix addressing boundary checks and type-safety issues uncovered during fuzz-testing.

#### PR Summary
This pull request improves SPDM command handling by tightening payload length validations and pointer arithmetic to prevent potential out-of-bound errors, while also refining error logging and memory cleanup routines.
- **`core/spdm/spdm_commands.c`**: Revised validations in SPDM negotiation and measurement responses by replacing offset-based checks with pointer comparisons and updated algorithm table traversal.
- **`core/spdm/spdm_commands.h`**: Updated the macro to perform pointer arithmetic using `uint8_t*` casts.
- **`core/spdm/cmd_interface_spdm.c`**: Enhanced error logging with explicit type casting for improved reliability.
- **`core/manifest/cfm/cfm_flash.c`**: Reset the `check_count` during manifest cleanup to ensure correct state reset.
- **`core/testing/spdm/spdm_commands_test.c`**: Adjusted the test case by setting the appropriate `payload_length`.

+++ core/attestation/attestation_requester.c
@@ -1458,7 +1458,7 @@ static int attestation_requester_process_spdm_protocol_response (
 				else {
 					debug_log_create_entry (DEBUG_LOG_SEVERITY_ERROR,
 						DEBUG_LOG_COMPONENT_ATTESTATION, ATTESTATION_LOGGING_SPDM_RESPONSE_ERROR,
-						((error_msg->error_code << 24) | (response->source_eid << 16) |
+						(((uint32_t) error_msg->error_code << 24) | (response->source_eid << 16) |
 									(response->target_eid << 8)), error_msg->error_data);
 					status = MSG_TRANSPORT_UNEXPECTED_RESPONSE;
 				}
